requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
apscheduler==3.10.4
python-dotenv==1.0.1
cinemagoer==2023.5.1
